Refatoração geral para aprimorar encapsulamento, legibilidade, testabilidade e aderência ao SRP
- Retirado lógica de négocio de dentro das classes das entidades

- Refatorado entidades para encapsular coleções e garantir o encapsulamento adequado

- Implementado Builders das entidades e Dtos

- Implementado o uso de MapStruct (package mapper)

- Refatoração da organização e estrutura do projeto para melhor consistência e compreensão

- Removido algumas constraints das tabelas do Postgre para se adequar as regras de negócio

- Adicionado testes de unidade para ProdutoService e SubCategoriaService

- Melhorado o código em geral

- Refatorado testes para usar os builders criados e melhor legibilidade

- Alterado chave primaria da entidade Pedido para UUID

- Alterado relacionamento de ItemPedido e Produto, agora ItemPedido tem apenas o id de Produto

+++ src/main/java/br/com/danielschiavo/shop/models/produto/arquivosproduto/ArquivoProdutoDTO.java
@@ -6,6 +6,6 @@ public record ArquivoProdutoDTO(
 		@NotNull
 		String nome,
 		@NotNull
-		Integer posicao) {
+		Byte posicao) {
 
 }
