+++ src/test/java/br/com/danielschiavo/shop/models/produto/validacoes/ValidadorPrimeiroArquivoImagemTest.java
@@ -12,8 +12,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import br.com.danielschiavo.shop.infra.exceptions.ValidacaoException;
-import br.com.danielschiavo.shop.models.produto.CadastrarProdutoDTO;
 import br.com.danielschiavo.shop.models.produto.arquivosproduto.ArquivoProdutoDTO;
+import br.com.danielschiavo.shop.models.produto.dto.CadastrarProdutoDTO;
 
 @ExtendWith(MockitoExtension.class)
 class ValidadorPrimeiroArquivoImagemTest {
@@ -27,7 +27,7 @@ class ValidadorPrimeiroArquivoImagemTest {
 	@Test
 	@DisplayName("Validador primeiro arquivo imagem não deve lançar exceção quando o arquivo na posição 0 realmente for uma imagem")
 	void ValidadorPrimeiroArquivoImagem_PrimeiroArquivoImagem_NaoDeveLancarExcecao() {
-		ArquivoProdutoDTO arquivoProdutoDTO = new ArquivoProdutoDTO("Imagem.jpeg", 0);
+		ArquivoProdutoDTO arquivoProdutoDTO = new ArquivoProdutoDTO("Imagem.jpeg", (byte) 0);
 		BDDMockito.given(cadastrarProdutoDTO.arquivos()).willReturn(List.of(arquivoProdutoDTO));
 		
 		Assertions.assertDoesNotThrow(() -> validador.validar(cadastrarProdutoDTO));
@@ -36,7 +36,7 @@ class ValidadorPrimeiroArquivoImagemTest {
 	@Test
 	@DisplayName("Validador primeiro arquivo imagem deve lançar exceção quando o arquivo na posição 0 não for uma imagem")
 	void ValidadorPrimeiroArquivoImagem_PrimeiroArquivoNaoEImagem_DeveLancarExcecao() {
-		ArquivoProdutoDTO arquivoProdutoDTO = new ArquivoProdutoDTO("Imagem.avi", 0);
+		ArquivoProdutoDTO arquivoProdutoDTO = new ArquivoProdutoDTO("Imagem.avi", (byte) 0);
 		BDDMockito.given(cadastrarProdutoDTO.arquivos()).willReturn(List.of(arquivoProdutoDTO));
 		
 		Assertions.assertThrows(ValidacaoException.class, () -> validador.validar(cadastrarProdutoDTO));
